+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,31 +1,31 @@
-package chess.board;
+package chess.domain.board;
 
-import static chess.position.InitialPositionFixtures.BLACK_BISHOP_LEFT_POSITION;
-import static chess.position.InitialPositionFixtures.BLACK_BISHOP_RIGHT_POSITION;
-import static chess.position.InitialPositionFixtures.BLACK_KING_POSITION;
-import static chess.position.InitialPositionFixtures.BLACK_KNIGHT_LEFT_POSITION;
-import static chess.position.InitialPositionFixtures.BLACK_KNIGHT_RIGHT_POSITION;
-import static chess.position.InitialPositionFixtures.BLACK_QUEEN_POSITION;
-import static chess.position.InitialPositionFixtures.BLACK_ROOK_LEFT_POSITION;
-import static chess.position.InitialPositionFixtures.BLACK_ROOK_RIGHT_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_BISHOP_LEFT_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_BISHOP_RIGHT_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_KING_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_KNIGHT_LEFT_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_KNIGHT_RIGHT_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_QUEEN_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_ROOK_LEFT_POSITION;
-import static chess.position.InitialPositionFixtures.WHITE_ROOK_RIGHT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_BISHOP_LEFT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_BISHOP_RIGHT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_KING_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_KNIGHT_LEFT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_KNIGHT_RIGHT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_QUEEN_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_ROOK_LEFT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.BLACK_ROOK_RIGHT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_BISHOP_LEFT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_BISHOP_RIGHT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_KING_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_KNIGHT_LEFT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_KNIGHT_RIGHT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_QUEEN_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_ROOK_LEFT_POSITION;
+import static chess.domain.position.InitialPositionFixtures.WHITE_ROOK_RIGHT_POSITION;
 
-import chess.piece.Bishop;
-import chess.piece.Color;
-import chess.piece.King;
-import chess.piece.Knight;
-import chess.piece.Pawn;
-import chess.piece.Piece;
-import chess.piece.Queen;
-import chess.piece.Rook;
-import chess.position.Position;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.Color;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
+import chess.domain.position.Position;
 import java.util.HashMap;
 import java.util.Map;
