+++ src/main/java/chess/domain/path/Movement.java
@@ -1,8 +1,8 @@
-package chess.path;
+package chess.domain.path;
 
-import chess.position.File;
-import chess.position.Position;
-import chess.position.Rank;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import java.util.Arrays;
 import java.util.NoSuchElementException;
