+++ src/main/java/chess/domain/position/File.java
@@ -1,41 +1,37 @@
-package chess.position;
+package chess.domain.position;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
 
-public enum Rank {
+public enum File {
 
-    ONE(1),
-    TWO(2),
-    THREE(3),
-    FOUR(4),
-    FIVE(5),
-    SIX(6),
-    SEVEN(7),
-    EIGHT(8);
+    A(1),
+    B(2),
+    C(3),
+    D(4),
+    E(5),
+    F(6),
+    G(7),
+    H(8);
 
     private final int value;
 
-    Rank(final int value) {
+    File(final int value) {
         this.value = value;
     }
 
-    static Rank from(int value) {
+    static File from(int value) {
         return Arrays.stream(values())
                 .filter(it -> it.value == value)
                 .findAny()
                 .orElseThrow(NoSuchElementException::new);
     }
 
-    int gapWith(Rank rank) {
-        return this.value - rank.value;
+    int gapWith(File file) {
+        return this.value - file.value;
     }
 
     public int value() {
         return value;
-    }
-
-    public boolean isSame(final int rank) {
-        return this.value == rank;
     }
 }
