+++ src/main/java/chess/domain/position/Rank.java
@@ -1,37 +1,41 @@
-package chess.position;
+package chess.domain.position;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
 
-public enum File {
+public enum Rank {
 
-    A(1),
-    B(2),
-    C(3),
-    D(4),
-    E(5),
-    F(6),
-    G(7),
-    H(8);
+    ONE(1),
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5),
+    SIX(6),
+    SEVEN(7),
+    EIGHT(8);
 
     private final int value;
 
-    File(final int value) {
+    Rank(final int value) {
         this.value = value;
     }
 
-    static File from(int value) {
+    static Rank from(int value) {
         return Arrays.stream(values())
                 .filter(it -> it.value == value)
                 .findAny()
                 .orElseThrow(NoSuchElementException::new);
     }
 
-    int gapWith(File file) {
-        return this.value - file.value;
+    int gapWith(Rank rank) {
+        return this.value - rank.value;
     }
 
     public int value() {
         return value;
+    }
+
+    public boolean isSame(final int rank) {
+        return this.value == rank;
     }
 }
