+++ src/test/java/chess/PathTest.java
@@ -2,8 +2,8 @@ package chess;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.path.Path;
-import chess.position.Position;
+import chess.domain.path.Path;
+import chess.domain.position.Position;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
